feat: genericize table schemas
