+++ solution/ui/regulations/eregs-component-lib/src/components/shared-components/Statutes/utils/enums.js
@@ -1,0 +1,7 @@
+const DISPLAY_TYPES = ["table", "list"];
+const TABLE_TYPES = ["ssa"];
+
+export {
+    DISPLAY_TYPES,
+    TABLE_TYPES,
+};
